MilkBot2.5/requirements.txt
PyNaCL
asyncio
discord.py
python-dotenv
youtube_dl
ffmpeg